+++ library/src/main/scala/minimal_algorithms/ranking/ExampleRanking.scala
@@ -11,7 +11,7 @@ object ExampleRanking {
     val input = spark.sparkContext.textFile(inputPath)
     val inputMapped = input.map(line => {
       val p = line.split(' ')
-      new ExampleMao(p(1).toInt)})
+      new ExampleMao(p(1).toDouble)})
 
     val minimalAlgorithm = new MinimalAlgorithm[ExampleMao](spark, 5)
     minimalAlgorithm.importObjects(inputMapped).computeRanking.saveAsTextFile(outputPath)
